pyproject.toml
[tool.poetry]
name = "room-booking"
version = "0.1.0"
description = "Room booking system at Innopolis University."
authors = ["Vladislav Deryabkin <vladeryabkin@gmail.com>"]
readme = "README.md"
repository = "https://github.com/inno-gang/room-booking"
packages = [{include = "room_booking"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.97.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
pyjwt = "^2.7.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pyright = "^1.1.314"
isort = "^5.12.0"

[tool.pyright]
include = ["app"]

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
